+++ App1/app1/Clases/Jornada.cs
@@ -7,14 +7,14 @@ namespace App1.Clases
 {
     class Jornada
     {
-        public Int64 Id { get; set; }
+        public int Id { get; set; }
         public string tipo_Jornada { get; set; }
         public string Franja_Horaria { get; set; }
 
           public Jornada()
         { }
 
-        public Jornada(Int64 pId, string pTipo_Jornada, string pFranja_Horaria)
+        public Jornada(int pId, string pTipo_Jornada, string pFranja_Horaria)
         {
             this.Id = pId;
             this.tipo_Jornada = pTipo_Jornada;
